requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
WTForms==2.3.3
Pillow==9.5.0
python-dotenv==0.19.2
Flask-Migrate==3.1.0
PyMySQL==1.0.2
cryptography==39.0.1
email-validator==1.1.3
gunicorn==20.1.0
